+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/like/articleLike/JpaArticleLikeRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.like.articleLike.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.like.articleLike;
 
 import com.mobble.mobbleserver.domain.like.articleLike.ArticleLike;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import java.util.List;
 import java.util.Optional;
 
-public interface ArticleLikeRepository extends JpaRepository<ArticleLike, Long> {
+public interface JpaArticleLikeRepository extends JpaRepository<ArticleLike, Long> {
 
     Optional<ArticleLike> findLikedByArticleIdAndMemberId(Long articleId, Long memberId);
