+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/like/clubLike/JpaClubLikeRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.like.clubLike.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.like.clubLike;
 
 import com.mobble.mobbleserver.domain.like.clubLike.ClubLike;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.Modifying;
 
 import java.util.Optional;
 
-public interface ClubLikeRepository extends JpaRepository<ClubLike, Long> {
+public interface JpaClubLikeRepository extends JpaRepository<ClubLike, Long> {
 
     Optional<ClubLike> findLikedByClubIdAndMemberId(Long clubId, Long memberId);
