+++ src/main/java/com/mobble/mobbleserver/infrastructure/persistence/like/commentLike/JpaCommentLikeRepository.java
@@ -1,4 +1,4 @@
-package com.mobble.mobbleserver.refactor.like.commentLike.repository;
+package com.mobble.mobbleserver.infrastructure.persistence.like.commentLike;
 
 import com.mobble.mobbleserver.domain.like.commentLike.CommentLike;
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.Modifying;
 import java.util.List;
 import java.util.Optional;
 
-public interface CommentLikeRepository extends CommentLikeQueryRepository, JpaRepository<CommentLike, Long> {
+public interface JpaCommentLikeRepository extends CommentLikeQueryRepository, JpaRepository<CommentLike, Long> {
 
     Optional<CommentLike> findLikedByCommentIdAndMemberId(Long commentId, Long memberId);
