+++ src/main/java/com/mobble/mobbleserver/refactor/like/articleLike/service/ArticleLikeQueryService.java
@@ -3,7 +3,7 @@ package com.mobble.mobbleserver.refactor.like.articleLike.service;
 import com.mobble.mobbleserver.refactor.article.entity.Article;
 import com.mobble.mobbleserver.refactor.article.validator.ArticleValidator;
 import com.mobble.mobbleserver.domain.like.articleLike.ArticleLike;
-import com.mobble.mobbleserver.refactor.like.articleLike.repository.ArticleLikeRepository;
+import com.mobble.mobbleserver.infrastructure.persistence.like.articleLike.JpaArticleLikeRepository;
 import com.mobble.mobbleserver.infrastructure.web.like.dto.response.LikeMemberListResponseDto;
 import com.mobble.mobbleserver.domain.like.baseLike.LikeType;
 import com.mobble.mobbleserver.refactor.like.baseLike.service.LikeQueryService;
@@ -19,7 +19,7 @@ import java.util.List;
 public class ArticleLikeQueryService implements LikeQueryService {
 
     private final ArticleValidator articleValidator;
-    private final ArticleLikeRepository articleLikeRepository;
+    private final JpaArticleLikeRepository articleLikeRepository;
 
     @Override
     public LikeType getType() {
